training/requirements/training.txt
datasets
polars[pyarrow]
torchmetrics